remove macos update phase (#227)
We should move the macos update step either to outside (so the tech can take care of it) - or perhaps make the script smarter when validating whether an update is needed. For now, this step causes a long delay when downloading macos updates that can be up to few gigabytes in size, and when apple servers time out, this breaks the rest of the build.